When estimating ship date, skip open ticketsks where actual > estimate.
This is a tough call, and I considered a number of options.  In the end,
I took the simplest---just ignre the tickets.

There is no solid way we can predict how many hours such a ticket
will take.  I considered looking at somehow using a developers velocity
history to build a PDF of an expected close date for this ticket, but
it didn't seem quite right to be using the velocity history to estimate
a ticket close date and then again to estimate the ship date.  Also,
the code would have been way more complex and the logic much hard to
explain in the manual.

Hopefully in practice, this is a case that won't occur that often.

What I should do is add a note to the developer summary table listing the
count of tickets skipped due to this condition, as well as the  tickets
skipped due to a zero estimate.